Fix: avoid underflow for very small widths

+++ CNucleus.h
@@ -262,7 +262,7 @@ class CNucleus : public CNuclide, public CWeight
   float LestoneCorrection(float Usaddle, float momInertiaEff,short iAfAn);
   static CRandom ran; //!< pointer to random number generator
   static float const pi; //!< 3.14159
-  static float const EkFraction; // !< calculates the Ek spectra down to this 
+  static double const EkFraction; // !< calculates the Ek spectra down to this
                                  // fraction of the maximum
   //functions
   CNucleus(int iZ,int iA);
